requirements.txt
streamlit==1.9.0
plotly_express==0.4.0